requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nltk==3.3
pyenchant==2.0.0
six==1.11.0
Werkzeug==0.14.1
gunicorn==19.9.0